Update docs install instructions (#620)
* Add note about Node version to install docs

I had enduring issues with the classic Gatsby sharp install errors
that weren't solved using the manual vips rebuild. I found this
error appeared in Node 16 / npm 8 (the engine specified in
package.json), and in 18+19, but at Node 20 / npm 10 it worked fine.

As deployment and various actions are only tested in v16 and this
was just an issue with local dev install, this change is only
recommended for local dev install until the Node version is
upgraded in the entire repo.

* Recommend --legacy-peer-deps npm install flag for local docs development

There's a mismatch of major Gatsby versions that causes install to fail
without this flag. @primer/gatsby-theme-doctocat@4.7.1 uses Gatsby 2 but
gatsby-plugin-postcss (which we use for CSS modules) requires Gatsby 3.