Add Playground story to ActionMenu
Because the props of ActionMenu are properly annotated, the Storybook
infers controls even when they're not intuitive for the playground -
for example: classname, ref, id. I've made the playground only show
controls for props we'd explicitly want to play with. Weirdly the
controls are filtered by their name rather than argType key which
is why they have to be hard-coded.

This also removes the args from the default story so the code example
is ready to use.