Reorder addon panel
This makes Controls the first panel in the addons so it opens
automatically (previously this was a11y). Since most people should
land on the Playground story for each component, defaulting to
Controls makes the most sense.

It then has Code, A11y, and Interactions. I removed the Actions addon
because we're not using it anywhere, it was just an empty panel in
the sidebar.

Ideally I'd like the Code addon to open when you open a Default,
feature, or example story, and the Controls to open when you open
a Playground story, but this doesn't seem to be possible with the SB
API.